requirements.txt
altair==5.5.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==44.0.3
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
filelock==3.19.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.3.2
packaging==24.2
pandas==2.3.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.0.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyogrio==0.11.1
pyOpenSSL==24.3.0
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.0
setuptools==78.1.1
shapely==2.1.1
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
stack-data==0.6.3
streamlit==1.40.2
streamlit-option-menu==0.4.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
umap==0.1.1
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
